style: add global styles; update ui comps
- add styles for tabs and button
- add new tabs size variant (sm, med)
- refactor user management page
