+++ calendar-ui/src/styles/globals.css
@@ -94,6 +94,12 @@
     --bc-blue-light: #1a4b82;
     --bc-gold: #fcba19;
     --bc-gold-dark: #e8a317;
+
+    /* Tabs underline (#0F6CBD equivalent) */
+    --tabs-underline: oklch(0.5263 0.1493 251.63);
+
+    /* Link / interactive blue (#0F6CBD equivalent) */
+    --link: oklch(0.5263 0.1493 251.63);
   }
 
   .dark {
@@ -130,6 +136,9 @@
     --sidebar-ring: oklch(0.439 0 0);
     --inverted-background: oklch(0.985 0 0);
     --inverted-foreground: oklch(0.145 0 0);
+
+    --tabs-underline: oklch(0.5263 0.1493 251.63);
+    --link: oklch(0.5263 0.1493 251.63);
   }
 }
 
@@ -155,6 +164,8 @@
   --color-card-foreground: var(--card-foreground);
   --color-inverted-background: var(--inverted-background);
   --color-inverted-foreground: var(--inverted-foreground);
+  --color-tabs-underline: var(--tabs-underline);
+  --color-link: var(--link);
   --radius-lg: var(--radius);
   --radius-md: calc(var(--radius) - 2px);
   --radius-sm: calc(var(--radius) - 4px);
@@ -224,6 +235,8 @@
   --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
   --color-sidebar-border: var(--sidebar-border);
   --color-sidebar-ring: var(--sidebar-ring);
+  --color-tabs-underline: var(--tabs-underline);
+  --color-link: var(--link);
 }
 
 :root {
@@ -259,6 +272,8 @@
   --sidebar-accent-foreground: oklch(0.205 0 0);
   --sidebar-border: oklch(0.922 0 0);
   --sidebar-ring: oklch(0.708 0 0);
+  --tabs-underline: oklch(0.5263 0.1493 251.63);
+  --link: oklch(0.5263 0.1493 251.63);
 }
 
 .dark {
@@ -293,6 +308,8 @@
   --sidebar-accent-foreground: oklch(0.985 0 0);
   --sidebar-border: oklch(1 0 0 / 10%);
   --sidebar-ring: oklch(0.556 0 0);
+  --tabs-underline: oklch(0.5263 0.1493 251.63);
+  --link: oklch(0.5263 0.1493 251.63);
 }
 
 @layer base {
